+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 
 ### Added
 
+### Changed
+
+- Support new runtime protocol session configuration
+
+### Added
+
 - Multi-character conversation support (2.0)
 
 ## [1.12.0] - 2024-05-06